Add files via upload
Input and Output, that is movement of the data across the different processes, pages or from the external sources.

To get the data/number from the main page or any above layer page to the below layer page.

In the below page, i.e., Page 1 here,  Click on Start -> Click Add in input -> Enter name, description, -> Drag drop data item in which to store in the value.
In the above layer page, i.e., Main Page here, click on Page Reference Stage -> In that details of input data item of the Page 1 will be already present. -> Just drag drop the data item from which the value has to be fetched.