Add files via upload
Simple Expiration Date Calculator, just to visualize the exchange of the data values between the pages. 
Addition of today's date with the number of days for trial, i.e,, 30 here.

The expiration date data element in the Main Page get the output from the ExpirationReminder Page through the ExpirationReminder Page Reference Stage element of the Main Page. 
In the ExpirationReminder Page we are calculating the today;s date with the number of days of the trial period. The number of days for the trial period we are getting from the main page as input and again sending the calculated expiration date as output to the main page.