Add files via upload
Using process reference stage to connect the different processes (earlier we were using the PAGE reference stage to connect different pages WITHIN the Process). This time we are connecting two or more different pages and using them as the subpages. 

And transfer the OUTPUTs between the subprocesses.

Master process (similar to the main page of a process) need to be created.